% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffon.R
\name{buffon}
\alias{buffon}
\title{Buffon's needle experiment}
\usage{
buffon(N, l = 1, d = 1)
}
\arguments{
\item{N}{number of needles to be thrown}

\item{l}{length of needle (default 1)}

\item{d}{distance between lines (default 1)}
}
\value{
Prints P(hit) and returns an estimate of pi
}
\description{
Simulates Buffon's experiment of throwing needles on a striped floor
}
\author{
Maria Frey & Cecilie Neupart
}
\keyword{buffon}
\keyword{carlo}
\keyword{experiment}
\keyword{monte}
\keyword{needle}
\keyword{simulation}

