% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens.R
\name{dens}
\alias{dens}
\title{Density}
\usage{
dens(x, d=NULL, h=NULL, method="naive")
}
\arguments{
\item{x}{a numeric variable}

\item{d}{the point to evaluate.}

\item{h}{the bandwidth. Will be calculated by Sturges for the naive method and Silverman's for the kernel method if not defined by user.}

\item{method}{the method to be used for the calculations. Must either be 'naive' (default) or the Gaussian 'kernel'.}
}
\value{
Returns the density in a single point d from data x using the specified method.\cr
 If the point d is not defined, it returns the densities at the "minimum, 1st Quantile, median, mean, 3rd Quantile and maximum of the data.
}
\description{
Estimates the density in a given point using the naive estimator or the Gaussian kernel.
}
\examples{
dens(faithful$eruptions)
}
\author{
Maria Frey & Cecilie Neupart
}
\keyword{density}
\keyword{estimate}
\keyword{gaussian}
\keyword{kernel}

