% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densplot.R
\name{densplot}
\alias{densplot}
\title{Density plot}
\usage{
densplot(x, n=500, method="naive", from=NULL, to=NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{number of points to be plotted}

\item{method}{the method to be used for the calculations. Must either be 'naive' (default) or the Gaussian 'kernel'.}

\item{from}{the point from which the plot should begin. Will default be one third of the standard deviation of x from min(x).}

\item{to}{the point where the plot should end. Will default be one third of the standard deviation of x from max(x).}
}
\value{
plots the estimated density function
}
\description{
Uses the dens function to plot the estimated density function of the dataset.
}
\author{
Maria Frey & Cecilie Neupart
}
\keyword{density}
\keyword{estimate}
\keyword{gaussian}
\keyword{kernel}
\keyword{plot}

