% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChain.R
\name{MarkovChain}
\alias{MarkovChain}
\title{Discrete Markov Chain}
\usage{
MarkovChain(p, k, n)
}
\arguments{
\item{p}{a transition probability matrix}

\item{k}{integer indicating the initial state of the chain}

\item{n}{number of simulated steps}
}
\value{
returns the exact and simulated stationary probabilities of each state.
}
\description{
Simulates the stationary distribution of a discrete Markov chain.
}
\examples{
MarkovChain(p = matrix(c(.1,.2,.7,0,.5,.5,.1,0,.9),ncol=3, byrow=TRUE), k = 2, n = 1000)
}
\author{
Maria Frey & Cecilie Neupart
}
\keyword{Markov}
\keyword{carlo}
\keyword{chain}
\keyword{discrete}
\keyword{monte}
\keyword{probabilities}
\keyword{stationary}

