% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regres.R
\name{regres}
\alias{regres}
\title{Linear Regression model}
\usage{
regres(formula)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}
}
\value{
Returns summary of coefficients, standard error, t-value, p-value and R^2
}
\description{
Fits linear models by regression, similar to lm.
}
\examples{
regres(faithful$eruptions ~ faithful$waiting)
}
\author{
Maria Frey & Cecilie Neupart
}
\keyword{linear}
\keyword{regression}

